Update eventlog configuration
+ PowerShell Profile Package Manager Init

+++ functions/Windows Event Log/Write-PPEventLog.ps1
@@ -8,7 +8,7 @@ $Metadata = @{
 	Author = "Janik von Rotz"
 	AuthorContact = "http://janikvonrotz.ch"
 	CreateDate = "2013-10-21"
-	LastEditDate = "2013-10-21"
+	LastEditDate = "2013-10-25"
 	Version = "1.0.0"
 	License = @'
 This work is licensed under the Creative Commons Attribution-ShareAlike 3.0 Switzerland License.
@@ -30,13 +30,13 @@ function Write-PPEventLog{
 .PARAMETER  Message
 	Message for the event log entry.
 
-.PARAMETER  Source
-	Source of the log event.
+.PARAMETER  EventLog
+	Name of the event log.
     
 .PARAMETER  EventID
 	Provide an EventId.
 
-.PARAMETER  EventLogSource
+.PARAMETER  Source
 	Source parameter for the event log entry.
 	
 .PARAMETER  EntryType
@@ -46,22 +46,22 @@ function Write-PPEventLog{
 	Appends the PowerShell session log to the message.
 
 .EXAMPLE
-	PS C:\> Write-PPEventLog -Message "User XY has not been notified" -Source $MyInvocation.InvocationName -EntryType Warning
+	PS C:\> Write-PPEventLog -Message "User XY has not been notified" -Source "Application XY" -EntryType Warning -EventId 123 -WriteMessage -AppendSessionLog
 #>
 
 	param(        
         [Parameter(Mandatory=$true)]
 		$Message,
-
-        [Parameter(Mandatory=$false)]
-		$Source,
         
         [Parameter(Mandatory=$false)]
-		$EventId,
+		$EventLog,
         
         [Parameter(Mandatory=$false)]
-		$EventLogSource,
-        
+		$EventId, 
+               
+        [Parameter(Mandatory=$false)]
+		$Source,       
+                
         [Parameter(Mandatory=$false)]
 		[string]
         $EntryType,
@@ -76,27 +76,44 @@ function Write-PPEventLog{
 	#--------------------------------------------------#
 	# main
 	#--------------------------------------------------#
-	if($WriteMessage){
-		if($EntryType -eq "Error"){Write-Error $Message    
-		}elseif($EntryType -eq "Warning"){Write-Warning $Message       
-		}else{Write-Host $Message}
+    $EventLogs =  Get-PPConfiguration $PSconfigs.EventLog.Filter | %{$_.Content.EventLog}
+    
+    if($EventLog){$EventLogs = $EventLogs | where{$_.Name -eq $EventLog}        
+    }elseif($EventLogs){$EventLogs = $EventLogs | where{$_.Role -eq "Default"} | select -first 1
+    }else{throw "Couldn't find event log configurations with role: Default in $($PSconfigs.Path)"}
+    
+    if($EventLogs -eq $null){throw "Couldn't find event log configurations in $($PSconfigs.Path)"}
+    
+    $EventLogs | %{   
+    	    
+        if(-not $EventId){    
+            if($EntryType -eq "Error"){$EventId = $_.ErrorEventId        
+            }elseif($EntryType -eq "Warning"){$EventId = $_.WarningEventId
+            }elseif($EntryType -eq "FailureAudit"){$EventId = $_.FailureAudit     
+            }elseif($EntryType -eq "SuccessAudit"){$EventId = $_.SuccessAudit
+            }elseif($EntryType -eq "InformationEventId"){$EventId = $_.InformationEventId     
+            }else{$EventId = $_.DefaultEventId}
+        }
+        
+        if(-not $EntryType){$EntryType = $_.DefaultEntryType}
+        
+        if($WriteMessage){
+    		if($EntryType -eq "Error"){Write-Error $Message    
+    		}elseif($EntryType -eq "Warning"){Write-Warning $Message       
+    		}else{Write-Host $Message}
+        }
+        
+        if(-not $Source){
+            $Source = $_.Source | where{$_.Role -eq "Default"} | %{"$($_.Name)"}
+            if($Source -eq $null){throw "Couldn't find default source in event log configuration for: $($_.Name)"}
+        }
+             
+        if($AppendSessionLog){
+    		Copy-Item $PSlogs.SessionFile ("$($PSlogs.SessionFile).tmp")
+    		$Message += "`n`n" + (Get-Content $PSlogs.SessionFile | Out-String)
+    		Remove-Item ("$($PSlogs.SessionFile).tmp")
+    	}       
+        
+        Write-EventLog -EventId $EventId -Message $Message -ComputerName $env:COMPUTERNAME -LogName $_.Name -Source $Source -EntryType $EntryType
     }
-	    
-    if(-not $EventId){
-    
-        if($EntryType -eq "Error"){$EventId = $PSlogs.ErrorEventId        
-        }elseif($EntryType -eq "Warning"){$EventId = $PSlogs.WarningEventId            
-        }else{$EventId = $PSlogs.InformationEventId}
-    }
-    
-	if($Source){$Message = $Source + "`n`n" + $Message}
-    if(-not $EventLogSource){$EventLogSource = "PowerShell Profile"}        
-    if($AppendSessionLog){
-		Copy-Item $PSlogs.SessionFile ("$($PSlogs.SessionFile).tmp")
-		$Message += "`n`n" + (Get-Content $PSlogs.SessionFile | Out-String)
-		Remove-Item ("$($PSlogs.SessionFile).tmp")
-	}
-    
-    if(-not $EntryType){ Write-EventLog -EventId $EventId -Message $Message -ComputerName $env:COMPUTERNAME -LogName $PSlogs.EventLogName -Source $EventLogSource
-    }else{Write-EventLog -EventId $EventId -Message $Message -EntryType $EntryType -ComputerName $env:COMPUTERNAME -LogName $PSlogs.EventLogName -Source $EventLogSource}
 }
